server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["nickAnhel <anhimovn1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
alembic = "^1.13.2"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
asyncpg = "^0.29.0"
aiobotocore = "^2.13.1"
aiofiles = "^24.1.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.3"
pillow = "^11.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
